+++ web/store/user.ts
@@ -49,6 +49,7 @@ export type UserState = {
   loading: boolean
   error?: string
   loggedIn: boolean
+  isPastPolicyGate?: boolean
   jwt: string
   profile?: AppSchema.Profile
   user?: AppSchema.User
@@ -261,6 +262,7 @@ function init(): UserState {
       loading: false,
       jwt: '',
       loggedIn: false,
+      isPastPolicyGate: false,
       ui,
       background,
       oaiUsageLoading: false,
@@ -272,6 +274,7 @@ function init(): UserState {
     loggedIn: true,
     loading: false,
     jwt: existing,
+    isPastPolicyGate: false,
     ui,
     background,
     oaiUsageLoading: false,
